refactor : 모델 패키지 세분화
lotto/lottoNumber, purchaseAmount, winningResult/Rank 로 세분화
winningNumber 패키지와 분리
