+++ src/main/java/lotto/dto/MatchingResultDto.java
@@ -1,4 +1,4 @@
-package lotto.model.dto;
+package lotto.dto;
 
 public record MatchingResultDto(int matchingAmount, boolean matchesBonusNumber) {
 }
