+++ src/main/java/lotto/model/lotto/winningResult/rank/RankDeterminer.java
@@ -1,4 +1,4 @@
-package lotto.model.rank;
+package lotto.model.lotto.winningResult.rank;
 
 public interface RankDeterminer<T> {
     Rank determine(T input);
