+++ src/main/java/lotto/service/winningNumber/NumberGenerator.java
@@ -1,4 +1,7 @@
-package lotto.model.winningNumber;
+package lotto.service.winningNumber;
+
+import lotto.model.winningNumber.BonusNumber;
+import lotto.model.winningNumber.WinningNumber;
 
 public interface NumberGenerator<T, U> {
     WinningNumber registerWinningNumber(T input);
