+++ src/test/java/lotto/model/lotto/LottosTest.java
@@ -6,6 +6,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import lotto.model.lotto.lottoNumber.Lotto;
+import lotto.model.lotto.lottoNumber.Lottos;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
