+++ src/test/java/lotto/model/lotto/PurchaseAmountTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import lotto.common.Exceptions;
-import lotto.model.purchaseAmount.PurchaseAmount;
+import lotto.model.lotto.purchaseAmount.PurchaseAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
