+++ src/test/java/lotto/model/rank/DefaultRankDeterminerTest.java
@@ -3,7 +3,9 @@ package lotto.model.rank;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
-import lotto.model.dto.MatchingResultDto;
+import lotto.dto.MatchingResultDto;
+import lotto.model.lotto.winningResult.rank.DefaultRankDeterminer;
+import lotto.model.lotto.winningResult.rank.Rank;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
