+++ src/test/java/lotto/model/winningNumber/DefaultBonusNumberGeneratorTest.java
@@ -6,7 +6,9 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import java.util.ArrayList;
 import java.util.Arrays;
 import lotto.common.Exceptions;
-import lotto.model.dto.AllWinningNumberDto;
+import lotto.dto.AllWinningNumberDto;
+import lotto.service.winningNumber.DefaultNumberGenerator;
+import lotto.service.winningNumber.NumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
