+++ src/test/java/lotto/model/winningNumber/DefaultWinningNumberGeneratorTest.java
@@ -6,6 +6,8 @@ import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 
 import java.util.List;
 import lotto.common.Exceptions;
+import lotto.service.winningNumber.DefaultNumberGenerator;
+import lotto.service.winningNumber.NumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
