+++ src/test/java/lotto/model/winningResult/WinningResultsTest.java
@@ -4,7 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.model.rank.Rank;
+import lotto.model.lotto.winningResult.WinningResult;
+import lotto.model.lotto.winningResult.WinningResults;
+import lotto.model.lotto.winningResult.rank.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
