+++ src/test/java/lotto/service/DefaultLottoMachineTest.java
@@ -5,15 +5,16 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import lotto.model.lotto.RandomNumberPicker;
-import lotto.model.rank.DefaultRankDeterminer;
+import lotto.model.lotto.lottoNumber.RandomNumberPicker;
+import lotto.model.lotto.winningResult.rank.DefaultRankDeterminer;
 import lotto.model.winningNumber.BonusNumber;
-import lotto.model.lotto.Lotto;
-import lotto.model.purchaseAmount.PurchaseAmount;
-import lotto.model.lotto.Lottos;
+import lotto.model.lotto.lottoNumber.Lotto;
+import lotto.model.lotto.purchaseAmount.PurchaseAmount;
+import lotto.model.lotto.lottoNumber.Lottos;
 import lotto.model.winningNumber.WinningNumber;
-import lotto.model.rank.Rank;
-import lotto.model.winningResult.WinningResults;
+import lotto.model.lotto.winningResult.rank.Rank;
+import lotto.model.lotto.winningResult.WinningResults;
+import lotto.service.lotto.DefaultLottoMachine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
